requirements.txt
# requirements.txt
asttokens==2.4.1
awscli==1.33.22
boto3==1.34.140
botocore==1.34.140
colorama==0.4.6
decorator==5.1.1
docutils==0.16
executing==2.0.1
ipython==8.26.0
jedi==0.19.1
jmespath==1.0.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.10.2
six==1.16.0
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.2.2
wcwidth==0.2.13
